refactor that from webcontroller to homecontroller and earthquakescontroller
